+++ grails-app/services/org/bbop/apollo/PermissionService.groovy
@@ -9,6 +9,7 @@ import org.apache.shiro.session.Session
 import org.apache.shiro.subject.Subject
 import org.bbop.apollo.gwt.shared.FeatureStringEnum
 import org.bbop.apollo.gwt.shared.PermissionEnum
+import org.bbop.apollo.preference.UserOrganismPreferenceDTO
 import org.codehaus.groovy.grails.web.json.JSONArray
 import org.codehaus.groovy.grails.web.json.JSONObject
 import org.codehaus.groovy.grails.web.servlet.mvc.GrailsParameterMap
@@ -63,7 +64,7 @@ class PermissionService {
             organismList.addAll(getOrganismsForGroup(userGroup))
         }
         List<Organism> returnOrganismList = []
-        for(Organism organism in organismList.sort(){ a,b -> a.commonName <=> b.commonName }){
+        for (Organism organism in organismList.sort() { a, b -> a.commonName <=> b.commonName }) {
             returnOrganismList.add(organism)
         }
 
@@ -145,7 +146,7 @@ class PermissionService {
 
         UserOrganismPermission userOrganismPermission = UserOrganismPermission.findByOrganismAndUser(organism, user)
         if (!userOrganismPermission) {
-            userOrganismPermission = new UserOrganismPermission(
+            new UserOrganismPermission(
                     organism: organism
                     , permissions: generatePermissionString(permissions)
                     , user: user
@@ -162,7 +163,7 @@ class PermissionService {
 
         GroupOrganismPermission groupOrganismPermission = GroupOrganismPermission.findByOrganismAndGroup(organism, group)
         if (!groupOrganismPermission) {
-            groupOrganismPermission = new GroupOrganismPermission(
+            new GroupOrganismPermission(
                     organism: organism
                     , permissions: generatePermissionString(permissions)
                     , group: group
@@ -174,7 +175,7 @@ class PermissionService {
         }
     }
 
-    private String generatePermissionString(List<PermissionEnum> permissionEnums) {
+    private static String generatePermissionString(List<PermissionEnum> permissionEnums) {
         JSONArray jsonArray = new JSONArray()
         for (PermissionEnum permissionEnum in permissionEnums) {
             jsonArray.add(permissionEnum.name())
@@ -250,7 +251,7 @@ class PermissionService {
         return []
     }
 
-    public static String getSequenceNameFromInput(JSONObject inputObject) {
+    static String getSequenceNameFromInput(JSONObject inputObject) {
         String trackName = null
         if (inputObject.has(FeatureStringEnum.SEQUENCE.value)) {
             trackName = inputObject.sequence
@@ -263,7 +264,7 @@ class PermissionService {
 
     // get current user from session or input object
     User getCurrentUser(JSONObject inputObject = new JSONObject()) {
-        String username
+        String username = null
         if (inputObject?.has(FeatureStringEnum.USERNAME.value)) {
             username = inputObject.getString(FeatureStringEnum.USERNAME.value)
         }
@@ -271,7 +272,7 @@ class PermissionService {
             username = SecurityUtils.subject.principal
         }
         if (!username) {
-            return null;
+            return null
         }
 
         User user = User.findByUsername(username)
@@ -304,7 +305,12 @@ class PermissionService {
         organism = getOrganismFromInput(inputObject)
 
         if (!organism) {
-            organism = preferenceService.getCurrentOrganismPreference(user, sequenceName, inputObject.getString(FeatureStringEnum.CLIENT_TOKEN.value))?.organism
+            String clientToken = inputObject.getString(FeatureStringEnum.CLIENT_TOKEN.value)
+            UserOrganismPreferenceDTO preferenceDTO = preferenceService.getCurrentOrganismPreference(user, sequenceName, clientToken)
+            log.debug "Permission service found DTO: ${preferenceDTO as JSON}"
+            if (preferenceDTO) {
+                organism = Organism.findById(preferenceDTO.organism.id)
+            }
         }
 
         Sequence sequence
@@ -338,7 +344,7 @@ class PermissionService {
         try {
             Session session = SecurityUtils.subject.getSession(false)
             if (session) {
-                Map<String, Integer> permissions = session.getAttribute(FeatureStringEnum.PERMISSIONS.getValue());
+                Map<String, Integer> permissions = (Map<String, Integer>) session.getAttribute(FeatureStringEnum.PERMISSIONS.getValue())
                 if (permissions) {
                     Integer permission = permissions.get(SecurityUtils.subject.principal)
                     PermissionEnum sessionPermissionsEnum = isAdmin() ? PermissionEnum.ADMINISTRATE : PermissionEnum.getValueForOldInteger(permission)
@@ -406,8 +412,9 @@ class PermissionService {
             }
             def authToken = new UsernamePasswordToken(jsonObject.username, jsonObject.password as String)
             try {
-                Subject subject = SecurityUtils.getSubject();
-                session = subject.getSession(true);
+                Subject subject = SecurityUtils.getSubject()
+                subject.getSession(true)
+//                session = subject.getSession(true)
 
                 subject.login(authToken)
                 if (!subject.authenticated) {
@@ -544,7 +551,7 @@ class PermissionService {
                     def req = request.JSON
                     def authToken = usernamePasswordToken ?: null
                     if (!authToken && req.username) {
-                        authToken = new UsernamePasswordToken(req.username, req.password)
+                        authToken = new UsernamePasswordToken(req.username as String, req.password as String)
                     }
                     if (authenticationService.authenticate(authToken, request)) {
                         log.info "Authenticated user ${authToken.username} using ${auth.name}"