+++ src/groovy/org/bbop/apollo/preference/OrganismDTO.groovy
@@ -1,0 +1,8 @@
+package org.bbop.apollo.preference
+
+class OrganismDTO {
+
+    Long id
+    String commonName
+    String directory
+}